fixed moon positioning, fixed some compiler warnings on macOS

+++ src/bit.c
@@ -16,7 +16,7 @@ char byte_to_char(uint8_t byte)
 int16_t bytes_to_int16_LE(uint8_t *buffer)
 {
     int16_t result = 0x0;
-    for (int i = 0; i < sizeof(int16_t); ++i)
+    for (unsigned long i = 0; i < sizeof(int16_t); ++i)
     {
         result = result | (int16_t)buffer[i] << (8 * i);
     }
@@ -26,7 +26,7 @@ int16_t bytes_to_int16_LE(uint8_t *buffer)
 int32_t bytes_to_int32_LE(uint8_t *buffer)
 {
     uint32_t result = 0x0;
-    for (int i = 0; i < sizeof(int32_t); ++i)
+    for (unsigned long i = 0; i < sizeof(int32_t); ++i)
     {
         result = result | (int32_t)buffer[i] << (8 * i);
     }
@@ -36,7 +36,7 @@ int32_t bytes_to_int32_LE(uint8_t *buffer)
 int64_t bytes_to_int64_LE(uint8_t *buffer)
 {
     uint64_t result = 0x0;
-    for (int i = 0; i < sizeof(int64_t); ++i)
+    for (unsigned long i = 0; i < sizeof(int64_t); ++i)
     {
         result = result | (int64_t)buffer[i] << (8 * i);
     }
@@ -48,7 +48,7 @@ int64_t bytes_to_int64_LE(uint8_t *buffer)
 uint16_t bytes_to_uint16_LE(uint8_t *buffer)
 {
     uint16_t result = 0x0;
-    for (int i = 0; i < sizeof(uint16_t); ++i)
+    for (unsigned long i = 0; i < sizeof(uint16_t); ++i)
     {
         result = result | (uint16_t)buffer[i] << (8 * i);
     }
@@ -58,7 +58,7 @@ uint16_t bytes_to_uint16_LE(uint8_t *buffer)
 uint32_t bytes_to_uint32_LE(uint8_t *buffer)
 {
     uint32_t result = 0x0;
-    for (int i = 0; i < sizeof(uint32_t); ++i)
+    for (unsigned long i = 0; i < sizeof(uint32_t); ++i)
     {
         result = result | (uint32_t)buffer[i] << (8 * i);
     }
@@ -68,7 +68,7 @@ uint32_t bytes_to_uint32_LE(uint8_t *buffer)
 uint64_t bytes_to_uint64_LE(uint8_t *buffer)
 {
     uint64_t result = 0x0;
-    for (int i = 0; i < sizeof(uint64_t); ++i)
+    for (unsigned long i = 0; i < sizeof(uint64_t); ++i)
     {
         result = result | (uint64_t)buffer[i] << (8 * i);
     }
